[test] reorganize zero/gemini tests
