+++ tests/test_moe/test_moe_colo_init.py
@@ -13,7 +13,7 @@ from colossalai.utils import free_port, get_current_device
 from colossalai.zero import ColoInitContext
 from tests.test_moe.test_moe_zero_init import MoeModel
 from tests.test_tensor.common_utils import debug_print
-from tests.test_zero.common import CONFIG
+from tests.test_zero.test_legacy.common import CONFIG
 
 
 @parameterize("init_device_type", ['cpu', 'cuda'])
