+++ tests/test_zero/test_legacy/test_gemini_manager.py
