+++ tests/test_zero/test_legacy/test_state_dict.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python
 # -*- encoding: utf-8 -*-
 
-from copy import deepcopy
 from functools import partial
 
 import pytest
